+++ src/main/java/domain/menu/Menus.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.menu;
 
 import java.util.ArrayList;
 import java.util.Collections;
