+++ src/main/java/domain/table/order/Count.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.table.order;
 
 public class Count {
