+++ src/main/java/domain/type/FunctionType.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.type;
 
 import java.util.Arrays;
