+++ src/main/java/domain/type/PaymentType.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.type;
 
 import java.util.Arrays;
 import java.util.function.Function;
