+++ src/test/java/domain/menu/MenuTest.java
@@ -1,4 +1,6 @@
-package domain;
+package domain.menu;
+
+import domain.type.Category;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
