+++ src/test/java/domain/table/TablesTest.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.table;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
