+++ src/test/java/domain/type/PaymentTypeTest.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.type;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
